Created mixin that allows a serializer to take 'fields' as an optional keyword argument.
