+++ server/utils/serializers.py
@@ -249,3 +249,16 @@ class ErrorSerializer(serializers.Serializer):
     error_code = serializers.IntegerField(allow_null=True, min_value=0, required=False)
     # Message for explanation of error. Usually a translated string.
     message = serializers.CharField(allow_null=True, allow_blank=True, required=False)
+
+
+class DynamicFieldsMixin:
+    """Mixin that allows a serializer to take 'fields' as an optional keyword argument."""
+    def __init__(self, *args, **kwargs):
+        fields = kwargs.pop('fields', None)
+        super().__init__(*args, **kwargs)
+
+        if fields is not None:
+            allowed = set(fields)
+            existing = set(self.fields)
+            for field_name in existing - allowed:
+                self.fields.pop(field_name)
